requirements-dev.txt
# Development and test dependencies
pytest
pre-commit
black
flake8
isort
mypy
pytest-cov
sphinx
pip-tools
markdown-it-py # Used to parse the readme to test the code examples

# Required for the game_example
pygame
